+++ serialporthandler.h
@@ -36,8 +36,8 @@ signals:
     void portNameChanged();
     void baudRateChanged();
     void newPortDetected(QString portName);
-    void newTemperatureData(int rollerNum, int widgetNum, float data);
-    void newAccelerometerData(int rollerNum, int widgetNum, float data);
+    void newTemperatureData(int rollerNum, int widgetNum, int data);
+    void newAccelerometerData(int rollerNum, int widgetNum, int data);
     void newRPMData(int rollerNum, int widgetNum, int data);
     void newLEDData(int rollerNum, int widgetNum, int data);
     void newMACAddressData(int rollerNum, QString macAddr);
